Split array api program gen into own file
So that we can see that its independent
