+++ python/tests/test_array_api.py
@@ -1,5 +1,5 @@
 from egglog.exp.array_api import *
-
+from egglog.exp.array_api_program_gen import *
 
 def test_simplify_any_unique():
     X = NDArray.var("X")
@@ -74,7 +74,7 @@ def test_sklearn_lda(snapshot_py):
     assume_shape(X_arr, TupleInt(150) + TupleInt(4))  # type: ignore
     assume_isfinite(X_arr)
 
-    y_arr = NDArray.var("y")
+    y_arr = NDArray.var("y") d
     assume_dtype(y_arr, int64)
     assume_shape(y_arr, TupleInt(150))  # type: ignore
     assume_value_one_of(y_arr, (0, 1, 2))  # type: ignore
